refactor: extract the map and pagination and hoist to client
